correcion repositorio

+++ CodigoFuente/antenas/API/DataSchema/ef_tipos_evento.cs
@@ -6,7 +6,7 @@ namespace API.DataSchema
 {
     public class ef_tipos_evento : IRegistroUnico
     {
-        public short id_tipo_evento { get; set; }
+        public int id_tipo_evento { get; set; }
         public string descripcion { get; set; } = null!;
         public bool activo { get; set; }
